third_party/WebKit/LayoutTests/fast/css/dynamic-style-attribute-query.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="testDiv">
</div>
<script>
document.getElementById('testDiv').style.top = 0;
shouldNotBe("document.querySelector('#testDiv[style]')", "null");
</script>
</body>
</html>
